feat: Update synchronized_at on oauth client on each sync
We want to update the synchronized_at on the io.cozy.oauth.clients every
time the extension is synced. It enables us to show this information in
cozy-settings.

To achieve that, I added a SyncService class that extends the jslib's
one and overrides its `setLastSync` method to call the
/settings/synchronized stack route. This class is imported when the
original one is imported. The webpack configs replaces the original one
with the override.