fix: CozyClient Usage
We can't import from cozy-client without making a deep import.
This is needed since we have a moduleResolution set to "node" in our
tsconfig for this project.

So it means that the main entry point from CozyClient is used, resulting
by importing "node stuff". It doesn't work.

I didn't want to take the risk to change the moduleResolution since
i'm afraid of the propagation.

Also, it seems that we added an alias for cozy-client in the webpack
configuration one day that use the node entry point also. I don't know
why but since we want to avoid passing by the "node" entry point, let's
remove it.