fix: handle flags in CozyClient service
Browser extension's code is split between background worker and UI
Popup

Objects that are instantiated on one context cannot be easily shared
with others. For example a service instantiated on the background
worker should not be used in the Popup's code

Flags components are living on the Popup's side. Then the
`CozyClientService` injected to those component has been instantiated
on Popup side

This is problematic because `CozyClientService` is instantiated on the
background worker side during the login phase

`CozyClientService` injected on Popup side is not synchronized with the
one on the background worker side (it is not even fully instantiated)
so we should rely as much as possible on the one from the background
worker

So this commit moves flags management on `CozyClientService` and Flags
components now request a flag status by using messaging services to
`CozyClientService`