Popup history fix for password and name generator
When editing a cipher, if the user went in the password or name generator, and then closed the popup, then the popup would not open correctly (only firefox). This was due to the "AddEditCipherInfo" expected but null when the popup was re-opened. The fix concists in injecting a fake cipher at history.service init with a "stateService.setAddEditCipherInfo"