fix: Use same CozyClientService instance in background and popup
CozyClientService was instanciated twice, one time in background and
one time in popup. In background, it stays instanciated when we close
the popup contrary to in popup.

We got issues when opening a Cozy link using getAppUrl() if we had
our lock setting to 'Never'. getAppUrl() *need* the client to be
created. So it is better to always use the background client which is
always created even when lock setting is 'Never'.