feat: For a Note, openWebApp opens notes in edit mode
If we are in the view component and our cipher is a paper of type
Note, we want to open notes app in edit mode instead of
passwords app.

We also set a returnUrl, so that when the user clicks on the back button on notes app, he is redirected to mespapiers app.