feat: Open new AddGeneric component from current-tab
In the `CurrentTab` component, the user can trigger an item creation
from multiple place

First one is the `+` button on the popup's to right corner. We want
that button to redirect to the new `AddGeneric` component

Second one is displayed when no cipher is found for current tab. Then
the Login section is replaced with a CTA section that would allow to
create a new Login for the current page. We want that button to skip
the new `AddGeneric` component and directly open the Cipher creation
with `Login` type