feat: Update generate tab icon
Same icon but with new way of adding it.

By using the .tab-icon-cozy class and a mask, we can change the font
icon by an SVG icon imported from cozy-ui and still update the color
of the icon when it is hovered.

It is better for maintenability because the icon is not hardcoded in
HTML.