feat: Add subtype and expirationData in Field
For expiration date custom fields, we want to display an expiration
annotation when the expiration date is expired or will expire soon.

But we do not want to add the expiration annotation directly in the
field value because this value is used
in the autofill and we do not want to have the expiration annotation
in the autofill.

So we extend Field (and related classes) with :
- a subtype
- an expiration data object that contains everything we need to
display the expiration annotation