feat: Apply correct theme on long based on the instance context
When the user log in a Cozy instance, we want to apply the correct
theme based on the instance configuration but also on the user's
preferences

The instance configuration is defined by the
`passwords.theme.default-contrasted`. If this flag is true, then the
default theme should be the `LightContrasted` one. Otherwise the
`System` theme should be applied

The user's preferences are defined when the user changes the app theme
from the settings page. From that point we want to keep the selected
theme even when the user logs out (this part is implemented in
following commit)

If the user did not change the app theme from the settings page, then a
log out should revert to the `LightContrasted` theme, which is the
app's default theme

Note that if the instance's flag is true, then we want to enforce the
`LightContrasted` whatever if the user did set a different theme or not
in their previous session