fix: Register CozyClient as an oauth login
Bitwarden extension is connected to cozy-stack with oauth2.
We added a CozyClient to the extension to fetch cozy data.
The CozyClient is based on the oauth2 tokens but was not
correctly registered as an oauth2 CozyClient.

Some CozyClient functions like getIconURL need to
know the login method to works properly,
so they were returning unusable results because
it behaved like if we were not in a oauth2 login.