feat: Add initials in contact view
We want to display a contact  avatar like in cozy-contacts. This avatar consists in a circle with contact initials and a background color depending of contact initials.

initials are calculated in cozy-client from the entire contact object. So we need to calculate initials when converting contact to cipher because we have the contact object only at this point.

We then add calculated initials to the contact view to use them later.