fix: Use tiny illustration thumbnail for paper page
Previously, we used tiny illustration thumbnail for paper list and medium illustration thumbnail for paper page.

The medium thumbnail is not generated automatically by the cozy-stack for PDF files contrary to the tiny thumbnail. So by default, the first access to the thumbnail was returning a default thumbnail. Switching to the tiny thumbnail is one solution among others (but the simplest).

The tiny thumbnail is of size 96x96. Previously we were displaying a thumbnail of 100px height (and potentially 100% width). So switching to the tiny thumbnail  (96px height) will have very small impact on the thumbnail displayed. Very small height and very big width thumbnails will be displayed smaller but they are not real use case of papers.

So we can switch to the tiny thumbnail.