feat: Add label to Field class
Label consists in :
- a type like "Perso" or "Mobile" or "Whatever you want"
- an optional label which can be "work" or "home"

For the moment we want to display the type next to the field name
(something like "Address - Perso"). We do not want to add the type
directly in the field name because the field name is used
for the autofill feature.

We will also need later the optional label for the autofill feature.