feat: Add id and parentId to Field class
By default, we want to display only the full address (i.e. "2 rue des pommiers 75000 Paris") as field. We want to display address parts (i.e. "2", "rue des pommiers", "75000", "Paris") as fields only if we click on the full address.

We start by adding an optional id and an optional parentId for fields. If a field has a parentId, we will display it on the view page only if the parentId is selected.