fix: Avatar color was not correct
ContactAvatarComponent is displayed with a *cdkVirtualFor loop. As
stated in the documentation, "To improve rendering performance,
*cdkVirtualFor caches previously created views after they are no
longer needed. When a new view would normally be created, a cached
view is reused instead.".

So we can not set the backgroundColor in ngOnInit because it will
keep the background of the first elements created.