fix: Do not return empty papers/contacts if only one paper/contact conversion fail
Previously, when converting papers or contacts to ciphers, we were
returning an empty array when an error occured. So with 1000
contacts, if one conversion failed, we were seeing 0 contact.

Now we will see the 999 contacts that were successfully converted and
we will log an error for each paper or contact that failed to be
converted.