feat: Reuse stored papers/contacts if sync when vault locked
We need the encryption key to convert papers/contacts. So if we try to do a sync when the vault is locked, we will have no papers/contacts. In some case, another sync is triggered when we unvault the lock, but if we unlock the vault when the previous sync is still in progress AND previous sync already passed the papers/contacts conversion (which failed), it can result to empty papers/contacts.

In this case, we want to reuse stored papers/contacts.