feat: Add cozyType to Field
To summarize, a Field has only a translated label and a value. It is meant to be displayed. But we used them extensively and distorted their initial purpose when adding papers and contacts.

Now we need to "keep the signification" of a Field to be able to use it in autofill feature. So instead of just having a label and a value like "Nom" and "Bob", we also want to keep a type like "name". This is the cozyType.