feat: Updating instead of deleting cache when deleting cipher
When deleting a cipher, we do not want to clear the cache because reconstructing the cache involves decrypting all ciphers which is very costly with our papers and contacts. Instead we manually update decrypted ciphers below.

Closing the cipher view after deleting a cipher was very slow before this commit.