fix: Use OnChanges instead of OnInit for cipher-row.component.ts
CipherRowComponent is displayed with a *cdkVirtualFor loop. As stated
in the documentation, "To improve rendering performance, *cdkVirtualFor
caches previously created views after they are no longer needed. When a
new view would normally be created, a cached view is reused instead."

So we can not set the isKonnector in ngOnInit because it will keep the
value of the first elements created

Related commit: 14338fea3817f0c407f415b5c7171d3ab7fa6575