feat: Avoid thumbnail blinking in cipher row
When a thumbnail URL is invalid, it triggers the onThumbnailError
handler and we display a default thumbnail in CSS. If the HTML view
is reloaded (which happens a lot), it triggers again and again the
onThumbnailError and the default thumbnail in CSS handler leading to
a blinking thumbnail.

Now we change directly the thumbnail URL with the default one. So no
onThumbnailError handler triggered. When the sync will be done, a new
cipher row will a new thumbnail URL.