fix: Try to autofill at best addresses created from cco2
Addresses created from cco2 have no context. cco2 filled what it
could fill in "formattedAddress" so we can have something like "rue
des Pommes, Paris" inside, but we have no idea of what is inside and
there is no "number" or "street" field.

So in this case, we fallback to returning to full "formattedAddress"
as the address even if it can contain the city.