feat: Reapply Cozy customization about autofill
In the previous merge upstream, we now we were going to use the new in page menu instead of the Cozy in page menu.

autofill.service.ts was a heavily modified file so I decided to choose Bitwarden merge upstream version.

So I reapply the relevant modification we need to be able to use the autofill feature for papers and contacts.