feat: Replace identities by contacts in inline menu
Each field in a form has a "filledByCipherType" value that represents if the inline menu should show logins, cards or identities.

Now for relevant form we set this "filledByCipherType" to CipherType.Contact instead of CipherType.Identity. It allows to disable very simply everything related to identities in the inline menu.

We also added the basics for contact support in inline autofill.