feat: Use new Cozy mapping to autofill basic contact data
To autofill a ContactView, we leverage what already exists for IdentityView.

Previsouly, we added a conversion function that created a IdentityView from a ContactView. Links between  Bitwarden and Cozy data was hardcoded.

Now we take benefit of the new mapping which is more reusable. We start with only basic contact data.