feat: Restore Cozy's navigation history feature
In #149 we implemented the History feature that allowed to restore the
popup's current page after closing and re-opening it

Since last merge upstream, we lost this feature due to the fact that
Bitwarden did rewrite their router implementation

The `UnauthGuardService` is not used anymore when opening the popup.
Instead `redirectGuard()` and `unauthGuardFn()` functions are used

Those methods don't have access to the HistoryService so we cannot copy
the old code into those new functions

To fix this, we want to create a new `RestoreHistory` component that
will be the new "default" route called when opening the popup. Then
this component will call the HistoryService to restore the correct
route from the history