fix: Update mocked ServerConfigResponse
Previously, Bitwarden added a server config fetched by a remote call.
We mocked it with a null value because we did not need more. Now,
we need a valid value inside "environment". So let's mock
everything to be sure everything works.

In a future cozy-stack release, cozy-stack will implement this route
and we will remove this mocked ServerConfigResponse.