feat: Use injected service in HistoryService
HistoryService was using BrowserApi.getBackgroundPage(); API which is
not available when using Manifest V3. Here we try by injecting
directly services in HistoryService. It was written that it was not
working, but now it seems to work.

I suspect that it is linked to the rewrite done recently on
HistoryService initialization.