feat: Init CozyClientService at beginning of popup
With Manifest v3, we get errors when opening Cozy links in the popup
because the client service is not instanciated.

In Manifest v2, we had no errors because the client service is
instanciated in the background page and we get it in the popup from
the background with getBgService.

In Manifest v3, it does not work the same. getBgService has been
adapted temporarly for Manifest v3 but :
- it is deprecated, we should migrate (how ? when ?)
- we need to initialize the client again

In this commit, I initialize the client again  in the popup. It does
not create a second oauth token because it reuse the same stored in
the background service worker.