refactor: Rename autofill options and variables to make them more generic
Previously, we added a CozyProfile option to the Bitwarden  autofill method to take into account the "Pro" or "Perso" profile when autofilling a phone for example. Variables were named in link with "ambiguous values".

But we will soon introduce the ability to autofill a paper from a list of paper which has no "Pro" or "Perso" profile and is not linked with "ambiguous values".

So here I renamed these options and variables to make them more generic.