fix: Hardcode tax revenue autofill for demande-logement-social.gouv.fr
On demande-logement-social.gouv.fr, we need to be able to autofill
the tax revenue for the two last years.

Previously, the two tax revenue fields had the target date like
"2022" in the "tag" field. But the form has changed and now, there is
no way to know the date of the tax revenue we want.

For the moment we hardcode the year targeted by each field with the
field ID.