fix: Do not autofill two times the same HTML field with contacts
The fill script is generated by taking into accounts :
- the cipher itself (login, card, contact)
- the fields of the cipher

For some reason, we added every contact field in the fields of the
cipher. So the autofill function sometimes try to autofill the same
HTML field two times : one because there is a match with the contact
and one because there is a match with the same value in fields.

We fix this by skipping fields during contact autofill.