feat: Do autofill when clicking on autofill buttons on popup
We customized Bitwarden autofill which now need a Cozy Client store
with all the contacts inside.

Now with Manifest V3, there are :
- an autofill service and a CozyClient service for the background
- an autofill service and a CozyClient service for the extension
popup

But only the CozyClient service of the background has its store
filled with all the contacts inside fetched during the fullSync.

So we send the autofill action from the extension popup to the
background.