feat: Set number or number + street in address field
In forms, we can have multiple possibilities for addresses :
- a field for the number (eg "1") and a field for the street (eg "rue
des champs")
- a field both (eg "1 rue des champs")

Here, we detect if there is a field for the number in the form to
know if we should autofill "rue des champs" or "1 rue des champs" in
the street field.

By modifying how we fill the street field, I also converted the
pathAttributes field of the model from a string[] to a string. It is
much simplier.