fix: Add a paper works again
Last implementation added support of addresses, but we lost the ability to know if we save a value to a contact or to a paper.

I changed the logic a little bit to support it again.

Kept :
- InputRef and InputRefValue duality (one for HTML element and one with the HTML element value)
- address is managed in a special case
- type and label are sent one time

Changed :
- we send the fieldQualifier with the HTML element value to be able to find the model.

So instead of sending :

{
  "street": "rue des champs"
}

we send  :

{
  "key": "street",
  "fieldQualifier": "identityAddress1", // allow to get the model
  "value": "rue des champs"
  }

Also added country in hidden fields.