feat: Display only papers with data in inline menu
It is possible to have papers like a "fiche de paie" without the data
associated in metadata.number because :
- the user did not fill it
- the konnector did not find the value

In this case we do not display the paper in the inline menu. We had a
quick fix in the selector field for the tax ref tax income. We
removed it and added a filter for every papers to check if the value
we want is present.