feat: Improve performance by limiting papers and contacts synced
Before we did the "fetch and convert papers" and the "fetch and
convert contacts" in parallel and we were keeping all of them. The
sync was very slow and the UI was very slow too if we had 1000+
elements.

Here we changed the process.

1. We fetch papers and contacts in parralel

2. We filter papers and contacts to keep only what we want. We need
both elements to filter them because papers and contacts are linked
(for example, we want to keep contacts that are referenced in
papers).

3. We convert them as ciphers

I also created a new sync.ts file to separate our code from the
Bitwarden sync service.

I also changed a little bit the queries to make them more efficient.