fix: Update contacts ciphers correctly when multiple updates
If on Cozy Contacts, I set 4 contacts as favorite in the same time, it will send 4 contact update events very quickly to the extension. The extension did not handle the events correctly so we could have race condition where we added only part of the favorite contacts to the extension.

So we add a synchronous queue that receive events and manage them sequentially.