fix: Remove Bitwarden wording
For some translations we previously added new translations with Cozy
wording. I removed them to update directly the Bitwarden translations
because it is safer. In the last merge upstream, Bitwarden used their
translation somewhere else where we did not see it so it is better to
change the translation directly.