fix: Open inline menu on focus if filled field only if settings is corresponding
We customized the inline menu to open it on focus even if the field
is filled. But we want to have this behavior only if the inline menu
visibility is set to "On focus". If the inline menu visibility is set
to "On button click" we want to keep the classic behavior.